feat: add async handlers support
