+++ src/main/ts/cli.ts
@@ -2,4 +2,13 @@
 
 import {run} from './index'
 
-run()
+let promise
+
+// tslint:disable-next-line:no-floating-promises
+(async() => {
+  promise = run()
+
+  await promise
+})()
+
+module.exports = promise
